refactor: move parse_repl_port/parse_repl_workspace_id into beamtalk-workspace
Both `beamtalk-mcp/src/workspace.rs` and `beamtalk-parity-tests/src/pool.rs`
contained identical logic for extracting the REPL port and workspace ID from
`beamtalk repl` stdout. pool.rs even had a comment acknowledging the copy.

Move the canonical implementations to `beamtalk-workspace` (which both crates
already depend on), rename to `parse_repl_workspace_id` for clarity, and
reduce the MCP module to thin delegating wrappers. Delete the private copies
and their duplicated tests from pool.rs; migrate the comprehensive test suite
to beamtalk-workspace.

Net change: -25 lines, one source of truth for both functions.
